worked on gun and on creating more realistic movement
I also creating my own Alex folder, added a script to the gun to fire a
raycast with an unfinished gizmos attached for the editor and I added a
second arm to the player that will theoretically sway back and forth
when the player is moving.
PS. It looks like I get more done when your not here.
JK